WordToPdf/requirements.txt
# 通用基础库
pandas>=1.3.0
openpyxl>=3.0.0
python-docx>=0.8.11
pillow>=8.0.0

# PDF相关
PyPDF2>=2.0.0
pdf2docx>=0.5.0
pdfplumber>=0.6.0
reportlab>=3.6.0
fpdf2>=2.5.0

# PPT相关
python-pptx>=0.6.0

# HTML相关
beautifulsoup4>=4.9.0
html5lib>=1.1

# 图片处理
img2pdf>=0.4.0
pdf2image>=1.16.0

# 构建相关
Cython>=0.29.0
numpy>=1.20.0
setuptools>=50.0.0

# 可选依赖
# docx2pdf>=0.1.8  # 需要系统安装Word